Improve the os_log and os_activity integration
