+++ Source/Base/ASLog.h
@@ -16,93 +16,101 @@
 //
 
 #import <AsyncDisplayKit/ASAvailability.h>
+#import <AsyncDisplayKit/ASBaseDefines.h>
 #import <Foundation/Foundation.h>
+#import <os/log.h>
+#import <os/activity.h>
 
-/// The signposts we use. Signposts are grouped by color. The SystemTrace.tracetemplate file
-/// should be kept up-to-date with these values.
-typedef NS_ENUM(uint32_t, ASSignpostName) {
-  // Collection/Table (Blue)
-  ASSignpostDataControllerBatch = 300,		// Alloc/layout nodes before collection update.
-  ASSignpostRangeControllerUpdate,				// Ranges update pass.
-  ASSignpostCollectionUpdate,							// Entire update process, from -endUpdates to [super perform…]
-
-  // Rendering (Green)
-  ASSignpostLayerDisplay = 325,						// Client display callout.
-  ASSignpostRunLoopQueueBatch,						// One batch of ASRunLoopQueue.
-
-  // Layout (Purple)
-  ASSignpostCalculateLayout = 350,				// Start of calculateLayoutThatFits to end. Max 1 per thread.
-
-  // Misc (Orange)
-  ASSignpostDeallocQueueDrain = 375,			// One chunk of dealloc queue work. arg0 is count.
-  ASSignpostCATransactionLayout,					// The CA transaction commit layout phase.
-  ASSignpostCATransactionCommit						// The CA transaction commit post-layout phase.
-};
-
-typedef NS_ENUM(uintptr_t, ASSignpostColor) {
-  ASSignpostColorBlue,
-  ASSignpostColorGreen,
-  ASSignpostColorPurple,
-  ASSignpostColorOrange,
-  ASSignpostColorRed,
-  ASSignpostColorDefault
-};
-
-static inline ASSignpostColor ASSignpostGetColor(ASSignpostName name, ASSignpostColor colorPref) {
-  if (colorPref == ASSignpostColorDefault) {
-    return (ASSignpostColor)((name / 25) % 4);
-  } else {
-    return colorPref;
-  }
-}
-
-#define ASMultiplexImageNodeLogDebug(...)
-#define ASMultiplexImageNodeCLogDebug(...)
-
-#define ASMultiplexImageNodeLogError(...)
-#define ASMultiplexImageNodeCLogError(...)
-
-#define AS_KDEBUG_ENABLE defined(PROFILE) && __has_include(<sys/kdebug_signpost.h>)
-
-#if AS_KDEBUG_ENABLE
-
-#import <sys/kdebug_signpost.h>
-
-// These definitions are required to build the backward-compatible kdebug trace
-// on the iOS 10 SDK.  The kdebug_trace function crashes if run on iOS 9 and earlier.
-// It's valuable to support trace signposts on iOS 9, because A5 devices don't support iOS 10.
-#ifndef DBG_MACH_CHUD
-#define DBG_MACH_CHUD 0x0A
-#define DBG_FUNC_NONE 0
-#define DBG_FUNC_START 1
-#define DBG_FUNC_END 2
-#define DBG_APPS 33
-#define SYS_kdebug_trace 180
-#define KDBG_CODE(Class, SubClass, code) (((Class & 0xff) << 24) | ((SubClass & 0xff) << 16) | ((code & 0x3fff)  << 2))
-#define APPSDBG_CODE(SubClass,code) KDBG_CODE(DBG_APPS, SubClass, code)
+#ifndef ASEnableLogs
+  #define ASEnableLogs 1
 #endif
 
-// Currently we'll reserve arg3.
-#define ASSignpost(name, identifier, arg2, color) \
-  AS_AT_LEAST_IOS10 ? kdebug_signpost(name, (uintptr_t)identifier, (uintptr_t)arg2, 0, ASSignpostGetColor(name, color)) \
-                    : syscall(SYS_kdebug_trace, APPSDBG_CODE(DBG_MACH_CHUD, name) | DBG_FUNC_NONE, (uintptr_t)identifier, (uintptr_t)arg2, 0, ASSignpostGetColor(name, color));
+#ifndef ASEnableVerboseLogging
+  #define ASEnableVerboseLogging 0
+#endif
 
-#define ASSignpostStartCustom(name, identifier, arg2) \
-  AS_AT_LEAST_IOS10 ? kdebug_signpost_start(name, (uintptr_t)identifier, (uintptr_t)arg2, 0, 0) \
-                    : syscall(SYS_kdebug_trace, APPSDBG_CODE(DBG_MACH_CHUD, name) | DBG_FUNC_START, (uintptr_t)identifier, (uintptr_t)arg2, 0, 0);
-#define ASSignpostStart(name) ASSignpostStartCustom(name, self, 0)
+ASDISPLAYNODE_EXTERN_C_BEGIN
 
-#define ASSignpostEndCustom(name, identifier, arg2, color) \
-  AS_AT_LEAST_IOS10 ? kdebug_signpost_end(name, (uintptr_t)identifier, (uintptr_t)arg2, 0, ASSignpostGetColor(name, color)) \
-                    : syscall(SYS_kdebug_trace, APPSDBG_CODE(DBG_MACH_CHUD, name) | DBG_FUNC_END, (uintptr_t)identifier, (uintptr_t)arg2, 0, ASSignpostGetColor(name, color));
-#define ASSignpostEnd(name) ASSignpostEndCustom(name, self, 0, ASSignpostColorDefault)
+/// Log for general node events e.g. interfaceState, didLoad.
+#define ASNodeLogEnabled 1
+os_log_t ASNodeLog();
+
+/// Log for layout-specific events e.g. calculateLayout.
+#define ASLayoutLogEnabled 1
+os_log_t ASLayoutLog();
+
+/// Log for display-specific events e.g. display queue batches.
+#define ASRenderLogEnabled 1
+os_log_t ASRenderLog();
+
+/// Log for collection events e.g. reloadData, performBatchUpdates.
+#define ASCollectionLogEnabled 1
+os_log_t ASCollectionLog();
+
+/// Log for ASNetworkImageNode and ASMultiplexImageNode events.
+#define ASImageLoadingLogEnabled 1
+os_log_t ASImageLoadingLog();
+
+/// Specialized log for our main thread deallocation trampoline.
+#define ASMainThreadDeallocationLogEnabled 0
+os_log_t ASMainThreadDeallocationLog();
+
+ASDISPLAYNODE_EXTERN_C_END
+
+/**
+ * The activity tracing system changed a lot between iOS 9 and 10.
+ * In iOS 10, the system was merged with logging and became much more powerful
+ * and adopted a new API.
+ *
+ * The legacy API is visible, but its functionality is extremely limited and the API is so different
+ * that we don't bother with it. For example, activities described by os_activity_start/end are not
+ * reflected in the log whereas activities described by the newer
+ * os_activity_scope are. So unfortunately we must use these iOS 10
+ * APIs to get meaningful logging data.
+ */
+#if OS_LOG_TARGET_HAS_10_12_FEATURES
+
+#define OS_ACTIVITY_NULLABLE                                              nullable
+#define AS_ACTIVITY_CURRENT                                               OS_ACTIVITY_CURRENT
+#define as_activity_scope(activity)                                       os_activity_scope(activity)
+#define as_activity_apply(activity, block)                                os_activity_apply(activity, block)
+#define as_activity_create(description, parent_activity, flags)           os_activity_create(description, parent_activity, flags)
+#define as_activity_scope_enter(activity, statePtr)                       os_activity_scope_enter(activity, statePtr)
+#define as_activity_scope_leave(statePtr)                                 os_activity_scope_leave(statePtr)
+#define as_activity_get_identifier(activity, outParentID)                 os_activity_get_identifier(activity, outParentID)
+
+#if ASEnableVerboseLogging
+  #define as_activity_scope_verbose(activity)                             as_activity_scope(activity)
+#else
+  #define as_activity_scope_verbose(activity)
+#endif
 
 #else
 
-#define ASSignpost(name, identifier, arg2, color)
-#define ASSignpostStartCustom(name, identifier, arg2)
-#define ASSignpostStart(name)
-#define ASSignpostEndCustom(name, identifier, arg2, color)
-#define ASSignpostEnd(name)
+#define OS_ACTIVITY_NULLABLE
+#define AS_ACTIVITY_CURRENT                                               OS_ACTIVITY_NULL
+#define as_activity_scope(activity)
+#define as_activity_scope_verbose(activity)
+#define as_activity_apply(activity, block)
+#define as_activity_create(description, parent_activity, flags)           OS_ACTIVITY_NULL
+#define as_activity_scope_enter(activity, statePtr)
+#define as_activity_scope_leave(statePtr)
+#define as_activity_get_identifier(activity, outParentID)                 (os_activity_id_t)0
 
+#endif // OS_LOG_TARGET_HAS_10_12_FEATURES
+
+/**
+ * The logging macros are not guarded by deployment-target checks like the activity macros are, but they are
+ * only available on iOS >= 9 at runtime, so just make them conditional.
+ */
+#define as_log_create(subsystem, category)  (AS_AT_LEAST_IOS9 ? os_log_create(subsystem, category) : (os_log_t)0)
+#define as_log_debug(log, format, ...)      (AS_AT_LEAST_IOS9 ? os_log_debug(log, format, ##__VA_ARGS__) : (void)0)
+#define as_log_info(log, format, ...)       (AS_AT_LEAST_IOS9 ? os_log_info(log, format, ##__VA_ARGS__) : (void)0)
+#define as_log_error(log, format, ...)      (AS_AT_LEAST_IOS9 ? os_log_error(log, format, ##__VA_ARGS__) : (void)0)
+#define as_log_fault(log, format, ...)      (AS_AT_LEAST_IOS9 ? os_log_fault(log, format, ##__VA_ARGS__) : (void)0)
+
+#if ASEnableVerboseLogging
+  #define as_log_verbose(log, format, ...)  as_log_debug(log, format, ##__VA_ARGS__)
+#else
+  #define as_log_verbose(log, format, ...)
 #endif
