+++ Source/Details/ASDataController.mm
@@ -24,8 +24,11 @@
 #import <AsyncDisplayKit/ASCollectionElement.h>
 #import <AsyncDisplayKit/ASCollectionLayoutContext.h>
 #import <AsyncDisplayKit/ASDispatch.h>
+#import <AsyncDisplayKit/ASDisplayNodeExtras.h>
 #import <AsyncDisplayKit/ASElementMap.h>
 #import <AsyncDisplayKit/ASLayout.h>
+#import <AsyncDisplayKit/ASLog.h>
+#import <AsyncDisplayKit/ASSignpost.h>
 #import <AsyncDisplayKit/ASMainSerialQueue.h>
 #import <AsyncDisplayKit/ASMutableElementMap.h>
 #import <AsyncDisplayKit/ASRangeManagingNode.h>
@@ -168,7 +171,11 @@ typedef void (^ASDataControllerCompletionBlock)(NSArray<ASCollectionElement *> *
   for (NSUInteger i = 0; i < count; i += batchSize) {
     NSRange batchedRange = NSMakeRange(i, MIN(count - i, batchSize));
     NSArray<ASCollectionElement *> *batchedElements = [elements subarrayWithRange:batchedRange];
-    NSArray<ASCellNode *> *nodes = [self _allocateNodesFromElements:batchedElements andLayout:shouldLayout];
+    NSArray<ASCellNode *> *nodes;
+    {
+      as_activity_scope(as_activity_create("Data controller batch", AS_ACTIVITY_CURRENT, OS_ACTIVITY_FLAG_DEFAULT));
+      nodes = [self _allocateNodesFromElements:batchedElements andLayout:shouldLayout];
+    }
     batchCompletionHandler(batchedElements, nodes);
   }
 
@@ -492,7 +499,14 @@ typedef void (^ASDataControllerCompletionBlock)(NSArray<ASCollectionElement *> *
   ASDisplayNodeAssertMainThread();
   
   if (changeSet.includesReloadData) {
-    _initialReloadDataHasBeenCalled = YES;
+    if (_initialReloadDataHasBeenCalled) {
+      as_log_debug(ASCollectionLog(), "reloadData %@", ASViewToDisplayNode(ASDynamicCast(self.dataSource, UIView)));
+    } else {
+      as_log_debug(ASCollectionLog(), "Initial reloadData %@", ASViewToDisplayNode(ASDynamicCast(self.dataSource, UIView)));
+      _initialReloadDataHasBeenCalled = YES;
+    }
+  } else {
+    as_log_debug(ASCollectionLog(), "performBatchUpdates %@ %@", ASViewToDisplayNode(ASDynamicCast(self.dataSource, UIView)), changeSet);
   }
   
   NSTimeInterval transactionQueueFlushDuration = 0.0f;
@@ -505,6 +519,7 @@ typedef void (^ASDataControllerCompletionBlock)(NSArray<ASCollectionElement *> *
   // See ASUICollectionViewTests.testThatIssuingAnUpdateBeforeInitialReloadIsUnacceptable
   // for the issue that UICollectionView has that we're choosing to workaround.
   if (!_initialReloadDataHasBeenCalled) {
+    as_log_debug(ASCollectionLog(), "%@ Skipped update because load hasn't happened.", ASObjectDescriptionMakeTiny(_dataSource));
     [changeSet executeCompletionHandlerWithFinished:YES];
     return;
   }
@@ -548,27 +563,36 @@ typedef void (^ASDataControllerCompletionBlock)(NSArray<ASCollectionElement *> *
   }
 
   // Mutable copy of current data.
-  ASElementMap *previousMap = _pendingMap;
-  ASMutableElementMap *mutableMap = [previousMap mutableCopy];
-  
-  BOOL canDelegateLayout = (_layoutDelegate != nil);
+  BOOL canDelegateLayout;
+  ASElementMap *newMap;
+  id layoutContext;
+  {
+    as_activity_scope(as_activity_create("Latch new data for collection update", changeSet.rootActivity, OS_ACTIVITY_FLAG_DEFAULT));
+    ASElementMap *previousMap = _pendingMap;
+    ASMutableElementMap *mutableMap = [previousMap mutableCopy];
 
-  // Step 1: Update the mutable copies to match the data source's state
-  [self _updateSectionContextsInMap:mutableMap changeSet:changeSet];
-  ASPrimitiveTraitCollection existingTraitCollection = [self.node primitiveTraitCollection];
-  [self _updateElementsInMap:mutableMap changeSet:changeSet traitCollection:existingTraitCollection shouldFetchSizeRanges:(! canDelegateLayout) previousMap:previousMap];
-  
-  // Step 2: Clone the new data
-  ASElementMap *newMap = [mutableMap copy];
-  self.pendingMap = newMap;
-  
-  // Step 3: Ask layout delegate for contexts
-  id layoutContext = nil;
-  if (canDelegateLayout) {
-    layoutContext = [_layoutDelegate layoutContextWithElements:newMap];
+    canDelegateLayout = (_layoutDelegate != nil);
+
+    // Step 1: Update the mutable copies to match the data source's state
+    [self _updateSectionContextsInMap:mutableMap changeSet:changeSet];
+    ASPrimitiveTraitCollection existingTraitCollection = [self.node primitiveTraitCollection];
+    [self _updateElementsInMap:mutableMap changeSet:changeSet traitCollection:existingTraitCollection shouldFetchSizeRanges:(! canDelegateLayout) previousMap:previousMap];
+
+    // Step 2: Clone the new data
+    newMap = [mutableMap copy];
+    self.pendingMap = newMap;
+
+    // Step 3: Ask layout delegate for contexts
+    if (canDelegateLayout) {
+      layoutContext = [_layoutDelegate layoutContextWithElements:newMap];
+    }
   }
+
+  as_log_debug(ASCollectionLog(), "New content: %@", newMap.smallDescription);
   
   dispatch_group_async(_editingTransactionGroup, _editingTransactionQueue, ^{
+    __block __unused os_activity_scope_state_s preparationScope = {}; // unused if deployment target < iOS10
+    as_activity_scope_enter(as_activity_create("Prepare nodes for collection update", AS_ACTIVITY_CURRENT, OS_ACTIVITY_FLAG_DEFAULT), &preparationScope);
     // Step 4: Allocate and layout elements if can't delegate
     NSArray<ASCollectionElement *> *elementsToProcess;
     if (canDelegateLayout) {
@@ -589,8 +613,9 @@ typedef void (^ASDataControllerCompletionBlock)(NSArray<ASCollectionElement *> *
       if (canDelegateLayout) {
         [_layoutDelegate prepareLayoutWithContext:layoutContext];
       }
-      
+
       [_mainSerialQueue performBlockOnMainThread:^{
+        as_activity_scope_leave(&preparationScope);
         [_delegate dataController:self willUpdateWithChangeSet:changeSet];
 
         // Step 5: Deploy the new data as "completed" and inform delegate
