+++ Source/Details/ASElementMap.m
@@ -183,6 +183,18 @@
   return [_elementToIndexPathMap countByEnumeratingWithState:state objects:buffer count:len];
 }
 
+- (NSString *)smallDescription
+{
+  NSMutableArray *sectionDescriptions = [NSMutableArray array];
+
+  NSUInteger i = 0;
+  for (NSArray *section in _sectionsOfItems) {
+    [sectionDescriptions addObject:[NSString stringWithFormat:@"<S%tu: %tu>", i, section.count]];
+    i++;
+  }
+  return ASObjectDescriptionMakeWithoutObject(@[ @{ @"itemCounts": sectionDescriptions }]);
+}
+
 #pragma mark - ASDescriptionProvider
 
 - (NSString *)description
