+++ Source/Details/ASObjectDescriptionHelpers.m
@@ -47,6 +47,9 @@ NSString *ASGetDescriptionValueString(id object)
       [strings addObject:[NSString stringWithFormat:@"%lu", (unsigned long)[indexPath indexAtPosition:i]]];
     }
     return [NSString stringWithFormat:@"(%@)", [strings componentsJoinedByString:@", "]];
+  } else if ([object respondsToSelector:@selector(componentsJoinedByString:)]) {
+    // e.g. "[ <MYObject: 0x00000000> <MYObject: 0xFFFFFFFF> ]"
+    return [NSString stringWithFormat:@"[ %@ ]", [object componentsJoinedByString:@" "]];
   }
   return [object description];
 }
@@ -56,7 +59,13 @@ NSString *_ASObjectDescriptionMakePropertyList(NSArray<NSDictionary *> * _Nullab
   NSMutableArray *components = [NSMutableArray array];
   for (NSDictionary *properties in propertyGroups) {
     [properties enumerateKeysAndObjectsUsingBlock:^(id  _Nonnull key, id  _Nonnull obj, BOOL * _Nonnull stop) {
-      [components addObject:[NSString stringWithFormat:@"%@ = %@", key, ASGetDescriptionValueString(obj)]];
+      NSString *str;
+      if (key == (id)kCFNull) {
+        str = ASGetDescriptionValueString(obj);
+      } else {
+        str = [NSString stringWithFormat:@"%@ = %@", key, ASGetDescriptionValueString(obj)];
+      }
+      [components addObject:str];
     }];
   }
   return [components componentsJoinedByString:@"; "];
@@ -73,7 +82,7 @@ NSString *ASObjectDescriptionMake(__autoreleasing id object, NSArray<NSDictionar
     return @"(null)";
   }
 
-  NSMutableString *str = [NSMutableString stringWithFormat:@"<%@: %p", [object class], object];
+  NSMutableString *str = [NSMutableString stringWithFormat:@"<%s: %p", object_getClassName(object), object];
 
   NSString *propList = _ASObjectDescriptionMakePropertyList(propertyGroups);
   if (propList.length > 0) {
