+++ Source/Layout/ASLayout.mm
@@ -316,9 +316,16 @@ static std::atomic_bool static_retainsSublayoutLayoutElements = ATOMIC_VAR_INIT(
 - (NSMutableArray <NSDictionary *> *)propertiesForDescription
 {
   NSMutableArray *result = [NSMutableArray array];
-  [result addObject:@{ @"layoutElement" : (self.layoutElement ?: (id)kCFNull) }];
-  [result addObject:@{ @"position" : [NSValue valueWithCGPoint:self.position] }];
   [result addObject:@{ @"size" : [NSValue valueWithCGSize:self.size] }];
+
+  if (auto layoutElement = self.layoutElement) {
+    [result addObject:@{ @"layoutElement" : layoutElement }];
+  }
+
+  auto pos = self.position;
+  if (!ASPointIsNull(pos)) {
+    [result addObject:@{ @"position" : [NSValue valueWithCGPoint:pos] }];
+  }
   return result;
 }
